### 2019-11-21 #1905 @jbloom
@will-rowe: Thanks so much for getting this started quickly. If you still want us to suggest additional reviewers, here are a few possibilities *not* on the JOSS list:
- [Martin Kircher](https://www.bihealth.org/en/research/research-groups/martin-kircher/)
- [Ben Murrell](https://ki.se/en/mtc/benjamin-murrell-research-project)

And a suggestion from the JOSS list is `afrubin`.
