### 2019-11-25 #1915 @whedon
[REVIEW]: alignparse: A Python package for parsing complex features from high-throughput long-read sequencing
**Submitting author:** @jbloom (<a href="http://orcid.org/0000-0003-1267-3408">Jesse Bloom</a>)
**Repository:** <a href="https://github.com/jbloomlab/alignparse" target ="_blank">https://github.com/jbloomlab/alignparse</a>
**Version:** 0.1.1
