+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Rhoads:2015,
-    author = {Rhoads, A. and Au K.F.},
+    author = {Rhoads, A. and Au, K.F.},
     title = {PacBio Sequencing and Its Applications},
     journal = {Genomics, Proteomics & Bioinformatics},
     year = 2015,
