### 2019-11-27 b78240c
added Genbank citation
