+++ paper/paper.bib
@@ -1,3 +1,13 @@
+@article{Sayers:2019,
+    author = {Sayers, E.W. and Cavanaugh, M. and Clark, K. and Ostell, J. and Pruitt, K.D. and Karsch-Mizrachi, I.},
+    title = {GenBank},
+    journal = {Nucleic Acids Research},
+    year = 2019,
+    month = October,
+    doi = {10.1093/nar/gkz956},
+    note = {gkz956},
+}
+
 @article{Rhoads:2015,
     author = {Rhoads, A. and Au, K.F.},
     title = {PacBio Sequencing and Its Applications},
