### 2019-11-27 8efbd9c
added full name of the Genbank format and a link to the NCBI docs
