### 2019-11-28 828dfe9
note that pysam needs extra dependencies to build htslib
