### 2019-11-28 996aa3a
clarified Python version requirement
