### 2019-11-28 63fa350
added basic compatibility information to README
