### 2019-11-28 jbloomlab/alignparse#53 @afrubin
YAML + FASTA or YAML only input files
After running the examples, it doesn't seem like there's any information in the Genbank files that couldn't be included in the YAML file. This would be more straightforward for users who are unfamiliar with the Genbank Flat File format and/or don't have the tools to generate these files.

Two options would be to continue to require two input files, but accept the sequence as a FASTA file instead of Genbank, or to allow the user to include the sequence in the YAML file itself.

This would be especially welcome because I didn't see anything in the `alignparse` documentation about how to generate the Genbank file. Some example code or a link to the relevant documentation elsewhere (Biopython?) should be added.
