### 2019-11-28 b0a2289
clarify what is meant by feature "accuracy"
