### 2019-11-28 f04c39c
explicitly mention that the examples are in the docs
