### 2019-11-28 a819df5
added suggestion for missing community guidelines
