### 2019-11-28 #1915 @afrubin
@jbloom @khdusenbury This is a really nice tool that I could see using for my own work. The manuscript is well written and I found the notebook-style documentation to be quite clear.

I've submitted a PR with some minor suggested edits to the text (https://github.com/jbloomlab/alignparse/pull/57) and opened four issues. Support for new input file formats requested in https://github.com/jbloomlab/alignparse/issues/53 should be considered optional and not required for acceptance.
