### 2019-12-02 jbloomlab/alignparse#58 @bede
PySAM problems with Python 3.8 + OS X
`pip install alignparse` fails for me inside a clean MacOS 10.14 + Python 3.8 environment created with conda due to a PySAM build issue. Using Python 3.7 or else installing pysam with conda in 3.8 works just fine.

Quite likely to be a Mac only issue that will be resolved in a future pysam update, but you could consider [creating an evironment.yml](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-from-an-environment-yml-file) to allow conda to users to install prebuilt dependencies via
`conda env create -f environment.yml`

Edit: conda would also allow managed installation of minimap2
