### 2019-12-02 #1915 @bede
A very well crafted tool and accompanying paper. All the bases have been well covered, and multiple use cases have been elegantly demonstrated through rich notebooks with lovely visualisations. The paper is clear and well written. I see that you have recently added contribution guidelines. @afrubin has already covered some other points.

I did encounter a minor installation issue that was easily worked around and have opened a corresponding issue https://github.com/jbloomlab/alignparse/issues/58. Actioning this is not required for acceptance. However, providing this conda installation option would also offer a route for users to simultaneously install a pinned and tested version of minimap2.
