### 2019-12-04 #1915 @will-rowe
Thank you @afrubin and @bede for your thorough and speedy reviews! Apologies for my slowness in getting back to this - I had to take some carer leave this week.

I think this is a lovely piece of work and we're close to acceptance. @jbloom - could you let us know if you intend to make any changes in light of the reviewer comments. I agree that support for new file formats or conda installation is not required for acceptance.
