### 2019-12-04 #1915 @khdusenbury
My apologies for being slow at getting the reviewer comments, but thank you all for the reviews and being so fast with getting to this!

I've started working on the issues @afrubin and @bede raised and hope to have their changes incorporated soon.
