### 2019-12-04 jbloomlab/alignparse#58 @khdusenbury
We have an `environment.yml` file in the main repo that I think should work for this. It specifies using Python 3.6 to avoid the issues with pysam. I just tested creating the environment locally on my machine and it seemed to work, but let me know if you run into issues! Thanks for your quick review, @bede!
