### 2019-12-04 4b20c6c
added testing and environment info to installation doc
