+++ docs/installation.rst
@@ -1,7 +1,8 @@
 Installation
 --------------
 
-``alignparse`` requires Python 3.6 or 3.7 and a Mac OS X or Linux operating system.
+``alignparse`` requires Python 3.6 or higher and is currently tested on Python 3.6 and 3.7. 
+A Mac OS X or Linux operating system is also required.
 
 The easiest way to install ``alignparse`` is from `PyPI <https://pypi.org/>`_ using `pip <https://pip.pypa.io>`_ with::
 
@@ -14,3 +15,7 @@ The source code for ``alignparse`` is available on GitHub at https://github.com/
 
 In order to use the :mod:`alignparse.minimap2` module, you need to install the `minimap2 <https://github.com/lh3/minimap2>`_ executable.
 You can do that via the `bioconda recipe <https://bioconda.github.io/recipes/minimap2/README.html>`_ or from the program's release page `as described here <https://github.com/lh3/minimap2#install>`_.
+
+An ``environment.yml`` file is provided with the source code. This can be used to create a `conda environment <https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html>`_ that installs ``alignparse`` and its dependencies using::
+
+    conda env create -f environment.yml
