### 2019-12-05 a256f92
added `minimap2` version requirements
