+++ docs/installation.rst
@@ -13,7 +13,7 @@ See the `pysam documentation <https://pysam.readthedocs.io/en/latest/installatio
 
 The source code for ``alignparse`` is available on GitHub at https://github.com/jbloomlab/alignparse.
 
-In order to use the :mod:`alignparse.minimap2` module, you need to install the `minimap2 <https://github.com/lh3/minimap2>`_ executable.
+In order to use the :mod:`alignparse.minimap2` module, you need to install the `minimap2 <https://github.com/lh3/minimap2>`_ executable (version 2.17 or higher).
 You can do that via the `bioconda recipe <https://bioconda.github.io/recipes/minimap2/README.html>`_ or from the program's release page `as described here <https://github.com/lh3/minimap2#install>`_.
 
 An ``environment.yml`` file is provided with the source code. This can be used to create a `conda environment <https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html>`_ that installs ``alignparse`` and its dependencies using::
