### 2019-12-05 5c902ab
Fixed `.query` bug.
For some reason, the df.query('status.str.match("^Success")') lines
started throwing 'Series' objects are not hashable errors.

Thus, I changed this splicing to explicitly splice on df['status'].
