### 2019-12-05 9bb6b4f
Edited paper to incorporate reviewer comments.
Also, re-ran lasv and flu example notebooks.
