+++ paper/paper.md
@@ -60,25 +60,28 @@ The [RecA deep mutational scanning library example](https://jbloomlab.github.io/
 Some viral genomes are sufficiently small to be sequenced in their entirety using long-read sequencing technology. 
 Recent work has shown that such long-read sequencing of viral genes can be combined with standard single-cell transcriptomic technologies (such as 10x Chromium) to simultaneously sequence the infecting virus and characterize the transcriptome in single infected cells [@Russell:2019]. 
 Such experiments require parsing the long-read viral sequences to identify viral mutations as well as cell barcodes, unique molecular identifiers, and other flanking sequences. 
-The [single-cell virus sequencing example](https://jbloomlab.github.io/alignparse/flu_virus_seq_example.html) shows how such parsing can readily be performed using [alignparse](https://jbloomlab.github.io/alignparse/). 
-
+The [single-cell virus sequencing example](https://jbloomlab.github.io/alignparse/flu_virus_seq_example.html) shows how such parsing can readily be performed using [alignparse](https://jbloomlab.github.io/alignparse/).
 
 # How alignparse works
 
 [alignparse](https://jbloomlab.github.io/alignparse/) takes the following inputs:
 
- 1. A user-defined Genbank file containing the sequence of one or more alignment targets with an arbitrary number of user-defined features.
+ 1. One or more user-defined Genbank files containing the sequence of one or more alignment targets with an arbitrary number of user-defined features. These Genbank files can be readily generated using sequence editing programs, such as [ApE](https://jorgensen.biology.utah.edu/wayned/ape/) or [Benchling](https://www.benchling.com/).
  
- 2. A YAML file containing parsing specifications for each feature. These specifications include filters indicating the maximal allowed mutations in each feature, as well as information on what output should be parsed for each feature (e.g., its sequence, its mutations, or simply if it is present)
+ 2. A YAML file containing parsing specifications for each feature. These specifications include filters indicating the maximal allowed mutations in each feature, as well as information on what output should be parsed for each feature (e.g., its sequence, its mutations, or simply if it is present).
 
-3. A FASTQ file containing the long-read sequencing data.
+3. A FASTQ file containing the long-read sequencing data. This file can be gzipped. There is no need to decompress gzipped FASTQ files first.
 
 These inputs are used to define a [Targets](https://jbloomlab.github.io/alignparse/alignparse.targets.html#alignparse.targets.Targets) object. 
 [alignparse](https://jbloomlab.github.io/alignparse/) then uses this [Targets](https://jbloomlab.github.io/alignparse/alignparse.targets.html#alignparse.targets.Targets) object to create sequence alignments and parse sequence features defined in the input Genbank and YAML files. 
 
 [alignparse](https://jbloomlab.github.io/alignparse/) aligns sequencing reads to the targets using [minimap2](https://github.com/lh3/minimap2).
 The [alignparse.minimap2](https://jbloomlab.github.io/alignparse/alignparse.minimap2.html) submodule provides alignment specifications optimized for the two example use cases described above. 
-[alignparse](https://jbloomlab.github.io/alignparse/) uses the [cs tags](https://lh3.github.io/minimap2/minimap2.html#10) generated by [minimap2](https://github.com/lh3/minimap2) to extract the relevant features from the alignments into intuitive data frames or CSV files.
+[alignparse](https://jbloomlab.github.io/alignparse/) uses the [cs tags](https://lh3.github.io/minimap2/minimap2.html#10) generated by [minimap2](https://github.com/lh3/minimap2) to extract the relevant features from the alignments into intuitive data frames or CSV files. 
+
[…]
