### 2019-12-05 8e747c3
add more info about about `seqsfileformat`
