+++ alignparse/targets.py
@@ -322,7 +322,13 @@ class Targets:
     allow_extra_features : bool
         Can targets have features not in `feature_parse_specs`?
     seqsfileformat : {'genbank'}
-        Format of `seqsfile`.
+        Format of `seqsfile`. Currently, 'genbank' is the only supported
+        option. More information about the required GenBank Flat File format
+        can be found at: https://www.ncbi.nlm.nih.gov/genbank/samplerecord/.
+        Genbank files can be readily generated using several sequence editing
+        programs, such as `ApE
+         <https://jorgensen.biology.utah.edu/wayned/ape/>`_ or `Benchling
+         <https://www.benchling.com/>`_.
     allow_clipped_muts_seqs : bool
         Returning sequence or mutations for features where non-zero
         clipping is allowed is dangerous, since as described in
