### 2019-12-05 6fb63a1
fix `seqsfileformat` description
