### 2019-12-05 f25fff8
edited `seqsfileformat` description
