### 2019-12-05 jbloomlab/alignparse#53 @khdusenbury
First, thank you for your thorough and quick review!

After looking into it a bit, I think support for different file formats for the input sequences may have to be a later feature. I added more info about the GenBank format and how to generate these files to both the paper and the documentation in the latest commits in my pull request. Hopefully, this will make the package easier to use for users that are less familiar with this file format.
