### 2019-12-06 86f21cc
updated `CHANGELOG.rst`
