+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
+0.1.1
+---------------------------
+Fixed DataFrame querying bug in `./alignparse/ccs.py`
+
 0.1.0
 ---------------------------
 Initial release
