### 2019-12-06 jbloomlab/alignparse#58 @khdusenbury
Implemented in PR #59.
