### 2019-12-06 jbloomlab/alignparse#53 @khdusenbury
Added additional info on GenBank format in PR #59.

Support for different file formats may be a future feature.
