### 2019-12-06 #1915 @khdusenbury
@will-rowe I think the manuscript and repo now incorporate @bede and @afrubin's reviews. Please let us know if there are additional changes you would like us to make. Thank you!
