### 2019-12-08 jbloomlab/alignparse#58 @afrubin
> @khdusenbury: Maybe we should also update the docs on installation to say that `alignparse` _requires_ Python 3.6 or higher, and is currently tested on Python 3.6 and 3.7. It should work on 3.8 (except for the `pysam` issues which are outside our purview), but right now the Travis tests only run for 3.6 and 3.7.

Just wanted to point out that the README still has the old language ("requires Python 3.6 or 3.7").
