### 2019-12-08 #1915 @afrubin
@will-rowe I've looked over the changes and updated manuscript and I'm happy to recommend that we move towards acceptance.
