### 2019-12-09 jbloomlab/alignparse#58 @jbloom
@afrubin: Good point. I've created a pull request for @khdusenbury to review that should address this issue (pull request #60).
