### 2019-12-10 #1915 @bede
The minor points I raised before have been addressed, and the tool now installs with a single command. Everything looks great from here.
