### 2019-12-11 #1915 @khdusenbury
@will-rowe Took me a bit to figure out Zenodo, but I think we have it archived properly. We archived version 0.1.1. and the DOI is 10.5281/zenodo.3569587. This is my first time working with Zenodo, so please let me know if there's anything I need to fix!
