### 2019-12-11 #1915 @will-rowe
Hey @openjournals/joss-eics - this is good to go. Thanks everyone
