### 2019-12-11 #1915 @kyleniemeyer
Hi @jbloom, I'm the Associate Editor-in-Chief on duty this week, so I'm going to do the final checks on your submission and then publish.

First off, can you edit the Zenodo entry metadata so that the title and authors match the submission? You can edit those without affecting the DOI.
