### 2019-12-11 #1915 @khdusenbury
@kyleniemeyer I think I successfully updated the author and title information in the latest PR. Thanks for handling the submission and please let me know if there are other changes to make.
