### 2019-12-11 #1915 @kyleniemeyer
@khdusenbury did you make the change at Zenodo? At https://doi.org/10.5281/zenodo.3569587 they look unchanged.
