### 2019-12-11 #1915 @khdusenbury
Whoops, I had just made the change in the repo. I think the changes should be on Zenodo now (and the repo `.zenodo.json` metadata file has been updated to match Zenodo).
