### 2019-12-11 #1915 @kyleniemeyer
Congrats @jbloom and @khdusenbury on your article's publication in JOSS!

Many thanks to @bede and @afrubin for reviewing, and @will-rowe for editing.
